requirements.txt
Django @ file:///tmp/build/80754af9/django_1605066529526/work
django-bootstrap3==14.2.0
django-braces==1.14.0
gunicorn==20.0.4
isort @ file:///tmp/build/80754af9/isort_1602603989581/work
lazy-object-proxy==1.4.3
numpy==1.20.1
opencv-python==4.5.1.48
Pillow==8.0.1
psycopg2 @ file:///C:/ci/psycopg2_1594305172475/work
psycopg2-binary==2.8.6
pycodestyle==2.6.0
pycparser==2.20
Pygments==2.6.1
pylint @ file:///C:/ci/pylint_1598617153160/work
pylint-django==2.3.0
pylint-plugin-utils==0.6
pytesseract==0.3.7
python-dateutil==2.8.1
pytz @ file:///tmp/build/80754af9/pytz_1606604771399/work
six @ file:///C:/ci/six_1605187374963/work
sqlparse @ file:///tmp/build/80754af9/sqlparse_1602184451250/work
tesseract==0.1.3
text-unidecode==1.3
toml @ file:///tmp/build/80754af9/toml_1592853716807/work
wincertstore==0.2
wrapt==1.11.2
